Better User Session Management
Added Logout button. Logout endpoint works. Logging out removes session key from db.
Added removeSessionKey method to dataService and userMapper to be used when logging out.
Completed sessionRetrieval endpoint to be called when the MainController is loaded, so that a user that did not log out before does not have to log back in.