Posts and Replies/Comments Refactor
Separated Posts and Comments (formerly known as Replies) into separate logic and objects.
Created Entry object which is extended by Post, Comment, and FullPost (Post with Comments).
Removed replyToPostId from Post and Post table.
Created a separate endpoint from adding Comments.
Added a working Comment box in the Post page which calls the new Add Comments endpoint.
Changed postTimestamp to timestamp.
Updated the CachingDataService to invalidate cache for a Post when a Comment is added to that Post.