Major Refactor and Clean Up
Switched the Configurations class from a singleton to an interface and an implementation class that can be instantiated.
Added a configuration for using the MockDataService.
Refactored the Setting and Settings classes, and added the Settings for List of Strings and List of Integers.
Changed JWTUtils to JWTService that can be instantiated, and moved the JWT Key into the class.
Routes -> RouteMappings.
Reorganized the pom file by adding properties for dependency versions.
Replaced the old jackson-xml-databind dependency for jackson-dataformat-xml and jackson-module-afterburner.