Edit and Delete API Are More Secure
The userId of the Post/Comment that is being edited or deleted is retrieved from the database and matched with the user making the request. Added a Map of Comment ID to Post ID to the CachingDataServices so that posts can be removed from the cache when the post's comments are added/edited/deleted.